pyproject.toml
[project]
name = "am-i-dying"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "pillow>=11.2.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
]
